Bump dependencies (#66)
* chore: bump lockfile (&& delete package-lock)

* doc: npm -> yarn

* chore: dependency bumps

* node-sass (node <= v14) -> sass (future proof)
* axios (major upgrade)

* chore: bump build script to 16

Explorer can now build on Node 16 (possibly 18). This was made possible by swapping deprecated package "node-sass" with "sass"

* chore: Bump tooling

Swapped deprecated vue-cli-service (Babbel + webpack based) for Vite (Esbuild). This is now the standard way of bootstrapping vue app

* fix: serves on port 8080, like in legacy codebase

* chore: remove webpack's "magic comments"

* chore: move index to root folder

* fix: inject build tag

* chore: dep bumps

* chore: rm aleph-js

Aleph-js has a ton of outdated dependency. It has a postinstall script for cryptographic primitives which is really time consuming (30s on avg), which are unused in the bundle.

* fix: build unknown bug

Vite "define" method does not use the same method under the hood when building or when inside  adevelopment server. It is not injected inside of the window "global" object when building (but is available in dev mode)